+++ src/slskd/Program.cs
@@ -752,27 +752,7 @@ namespace slskd
 
         private static void InitSQLiteOrFailFast()
         {
-            // initialize
-            // avoids: System.Exception: You need to call SQLitePCL.raw.SetProvider().  If you are using a bundle package, this is done by calling SQLitePCL.Batteries.Init().
-            SQLitePCL.Batteries.Init();
-
-            // check the threading mode set at compile time. if it is 0 it is unsafe to use in a multithreaded application, which slskd is.
-            // https://www.sqlite.org/compile.html#threadsafe
-            var threadSafe = SQLitePCL.raw.sqlite3_threadsafe();
-
-            if (threadSafe == 0)
-            {
-                throw new InvalidOperationException($"SQLite binary was not compiled with THREADSAFE={threadSafe}, which is not compatible with this application. Please create a GitHub issue to report this and include details about your environment.");
-            }
-
-            Log.Debug("SQLite was compiled with THREADSAFE={Mode}", threadSafe);
-
-            if (SQLitePCL.raw.sqlite3_config(SQLitePCL.raw.SQLITE_CONFIG_SERIALIZED) != SQLitePCL.raw.SQLITE_OK)
-            {
-                throw new InvalidOperationException($"SQLite threading mode could not be set to SERIALIZED ({SQLitePCL.raw.SQLITE_CONFIG_SERIALIZED}). Please create a GitHub issue to report this and include details about your environment.");
-            }
-
-            Log.Debug("SQLite threading mode set to {Mode} ({Number})", "SERIALIZED", SQLitePCL.raw.SQLITE_CONFIG_SERIALIZED);
+            StartupSqlite.InitOrFailFast(Log);
         }
 
         private static void ConfigureGlobalLogger()